cgiar-crop-identification/requirements.txt
ipykernel==6.25.2
ipython==8.16.1
joblib==1.3.2
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyterlab-widgets==3.0.9
matplotlib==3.8.0
matplotlib-inline==0.1.6
numpy==1.26.1
opencv-python==4.8.1.78
pandas==2.1.1
Pillow==10.1.0
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
timm==0.9.8
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.34.1
xgboost==2.0.0
